added price count functionallity (pivot)

+++ app/Http/Controllers/CartController.php
@@ -28,19 +28,35 @@ class CartController extends Controller
   } else {
     $order = Order::find($orderId);
   }
-$order->products()->attach($productId);
+  if($order->products->contains($productId)){
+    $pivotRow = $order -> products()->where('product_id',$productId)->first();
+    $pivotRow->count++;
+    $pivotRow->update();
+  } else {
+    $order->products()->attach($productId);
+  }
     return redirect()->route('cart');
   }
 
   public function cartRemove($productId)
   {
-    $orderId = session('orderId');
-    if (is_null($orderId)){
+    $orderid = session('orderId');
+    if (is_null($orderId))
+    {
       return redirect()->route('cart');
     }
-      $order = Order::find('$ordedId');
-      $order->products()->detach($productId);
-      return redirect()->route('cart');
+    $order = Order::find($orderId);
+
+    if($order->products->contains($productId)) {
+      $pivotRow=$order->products()->where('product_id',$productId)->first()->pivot;
+      if($pivotRow->count<2){
+        $order->products()->detach($productId);
+      } else {
+        $pivotRow->count--;
+        $pivotRow->update();
+      }
+    }
+    return redirect()->route('cart');
   }
